# finalproject/flask-project/models/static/swagger.yaml
openapi: 3.0.0
info:
  title: Listening to music
  version: 1.0.0
  description: This is description about a music.
paths:
...
  /song/{type}:
    put:
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            example: Blueberry Faygo
      responses:
        '200':
          description: The song has been update.
        '500':
          $ref: '#/components/responses/500ApiError'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/song'
  /song/{type}/{title}:
    post:
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            example: Blueberry Faygo
        - in: path
          name: title
          required: true
          schema:
            type: string
            example: Blueberry Faygo
      description: Add a new music 
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/song'
          application/xml:
              schema:
                $ref: '#/components/schemas/song'
      responses:
        '200':
          description: The music has been created.
        '500':
          $ref: '#/components/responses/500ApiError'
...
components:
  schemas:
    song:
      type: object
      properties:
        title:
          type: string
          example: Life Goes On
        artist:
          type: string
          example: Post Malone
        album:
          type: string
          example: Stoney
        year:
          type: integer
          example: 2020
        country:
          type: string
          example: American
        genre:
          type: string
          example: Hip-Hop/Rap
        web_url:
          type: string
          example: https://genius.com/Post-malone-go-flex-lyrics
        img_url:
          type: string
          example: https://i.ytimg.com/vi/tQjsAJhsSw8/maxresdefault.jpg
        yt_url:
          type: string
          example: tQjsAJhsSw8
       
      required: 
        - title
...
  responses:
...
    500ApiError:
      description: This is unexpected
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: string
                example: 500
              message:
                type: string
                example: This is an error
...